author agent implementation
